Add tests for WV visual-customizations
